+++ Port-win32/ClntService.cpp
@@ -4,7 +4,7 @@
  * authors: Tomasz Mrugalski <thomson@klub.com.pl>
  *          Marek Senderski <msend@o2.pl>
  *
- * $Id: ClntService.cpp,v 1.21 2005-07-24 16:00:02 thomson Exp $
+ * $Id: ClntService.cpp,v 1.22 2005-08-07 18:10:59 thomson Exp $
  *
  * Released under GNU GPL v2 licence
  *
@@ -51,7 +51,7 @@ EServiceState TClntService::ParseStandardArgs(int argc,char* argv[])
 			}
 			n++;
 			char temp[255];
-			int endpos;
+			size_t endpos;
 			strncpy(temp,argv[n],255);
 			endpos = strlen(temp);
 			if (temp[endpos-1]=='\\')
@@ -130,6 +130,9 @@ TClntService::~TClntService(void)
 
 /*
  * $Log: not supported by cvs2svn $
+ * Revision 1.21  2005/07/24 16:00:02  thomson
+ * Port WinNT/2000 related changes.
+ *
  * Revision 1.20  2005/07/17 21:09:52  thomson
  * Minor improvements for 0.4.1 release.
  *
